Moved losses and layers out of utils
