+++ src/mlsquare/imly/architectures/sklearn.py
@@ -8,8 +8,9 @@ from keras.models import Model
 from sklearn.preprocessing import OneHotEncoder
 from ..base import registry, BaseModel
 from ..adapters.sklearn import SklearnKerasClassifier, SklearnKerasRegressor
-from ..utils.layers import DecisionTree
+from ..layers.keras import DecisionTree
 from ..utils.functions import _parse_params
+from ..losses import lda_loss
 
 
 class GeneralizedLinearModel(BaseModel):  # Rename glm
